+++ xarray/tests/test_weighted.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from typing import Any, Iterable
+
 import numpy as np
 import pytest
 
@@ -654,23 +656,49 @@ def test_weighted_quantile_3D(dim, q, add_nans, skipna):
     assert_allclose(expected, result2.data)
 
 
-def test_weighted_operations_nonequal_coords():
-    # There are no weights for a == 4, so that data point is ignored.
-    weights = DataArray(np.random.randn(4), dims=("a",), coords=dict(a=[0, 1, 2, 3]))
-    data = DataArray(np.random.randn(4), dims=("a",), coords=dict(a=[1, 2, 3, 4]))
-    check_weighted_operations(data, weights, dim="a", skipna=None)
+@pytest.mark.parametrize(
+    "coords_weights, coords_data, expected_value_at_weighted_quantile",
+    [
+        ([0, 1, 2, 3], [1, 2, 3, 4], 2.5),  # no weights for coord a == 4
+        ([0, 1, 2, 3], [2, 3, 4, 5], 1.8),  # no weights for coord a == 4 or 5
+        ([2, 3, 4, 5], [0, 1, 2, 3], 3.8),  # no weights for coord a == 0 or 1
+    ],
+)
+def test_weighted_operations_nonequal_coords(
+    coords_weights: Iterable[Any],
+    coords_data: Iterable[Any],
+    expected_value_at_weighted_quantile: float,
+) -> None:
+    """Check that weighted operations work with unequal coords.
 
-    q = 0.5
-    result = data.weighted(weights).quantile(q, dim="a")
-    # Expected value computed using code from https://aakinshin.net/posts/weighted-quantiles/ with values at a=1,2,3
-    expected = DataArray([0.9308707], coords={"quantile": [q]}).squeeze()
-    assert_allclose(result, expected)
 
-    data = data.to_dataset(name="data")
-    check_weighted_operations(data, weights, dim="a", skipna=None)
+    Parameters
+    ----------
+    coords_weights : Iterable[Any]
+        The coords for the weights.
+    coords_data : Iterable[Any]
+        The coords for the data.
+    expected_value_at_weighted_quantile : float
+        The expected value for the quantile of the weighted data.
+    """
+    da_weights = DataArray(
+        [0.5, 1.0, 1.0, 2.0], dims=("a",), coords=dict(a=coords_weights)
+    )
+    da_data = DataArray([1, 2, 3, 4], dims=("a",), coords=dict(a=coords_data))
+    check_weighted_operations(da_data, da_weights, dim="a", skipna=None)
 
-    result = data.weighted(weights).quantile(q, dim="a")
-    assert_allclose(result, expected.to_dataset(name="data"))
+    quantile = 0.5
+    da_actual = da_data.weighted(da_weights).quantile(quantile, dim="a")
+    da_expected = DataArray(
+        [expected_value_at_weighted_quantile], coords={"quantile": [quantile]}
+    ).squeeze()
+    assert_allclose(da_actual, da_expected)
+
+    ds_data = da_data.to_dataset(name="data")
+    check_weighted_operations(ds_data, da_weights, dim="a", skipna=None)
+
+    ds_actual = ds_data.weighted(da_weights).quantile(quantile, dim="a")
+    assert_allclose(ds_actual, da_expected.to_dataset(name="data"))
 
 
 @pytest.mark.parametrize("shape_data", ((4,), (4, 4), (4, 4, 4)))