+++ doc/whats-new.rst
@@ -55,6 +55,8 @@ Bug fixes
   By `Fabian Hofmann <https://github.com/FabianHofmann>`_.
 - Fix step plots with ``hue`` arg. (:pull:`6944`)
   By `András Gunyhó <https://github.com/mgunyho>`_.
+- Avoid use of random numbers in `test_weighted.test_weighted_operations_nonequal_coords` (:issue:`6504`, :pull:`6961`).
+  By `Luke Conibear <https://github.com/lukeconibear>`_.
 
 Documentation
 ~~~~~~~~~~~~~